feat: Added exception handling : 'id not found'

+++ src/main/java/com/techreturners/bookmanager/service/BookManagerServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class BookManagerServiceImpl implements BookManagerService {
@@ -15,9 +16,11 @@ public class BookManagerServiceImpl implements BookManagerService {
     BookManagerRepository bookManagerRepository;
 
     @Override
-    public List<Book> getAllBooks() {
+    public List<Book> getAllBooks() throws RecordNotFoundException {
         List<Book> books = new ArrayList<>();
         bookManagerRepository.findAll().forEach(books::add);
+        if (!books.isEmpty())
+            throw new RecordNotFoundException("Book Records Not Found !!!!");
         return books;
     }
 
@@ -27,15 +30,21 @@ public class BookManagerServiceImpl implements BookManagerService {
     }
 
     @Override
-    public Book getBookById(Long id) {
-        return bookManagerRepository.findById(id).get();
+    public Book getBookById(Long id) throws RecordNotFoundException {
+
+        Optional<Book> book = bookManagerRepository.findById(id);
+        if (!book.isPresent())
+            throw new RecordNotFoundException("Book with id : " + id + " not Found in DB !!!");
+        return book.get();
     }
 
     //User Story 4 - Update Book By Id Solution
     @Override
-    public void updateBookById(Long id, Book book) {
+    public void updateBookById(Long id, Book book) throws RecordNotFoundException  {
+        Optional<Book> getBook = bookManagerRepository.findById(id);
+        if (!getBook.isPresent())
+            throw new RecordNotFoundException("Update failed !!! Book with id : " + id + " not Found in DB !!!");
         Book retrievedBook = bookManagerRepository.findById(id).get();
-
         retrievedBook.setTitle(book.getTitle());
         retrievedBook.setDescription(book.getDescription());
         retrievedBook.setAuthor(book.getAuthor());
@@ -45,7 +54,10 @@ public class BookManagerServiceImpl implements BookManagerService {
     }
 
     @Override
-    public void deleteBookById(Long bookId) {
+    public void deleteBookById(Long bookId) throws RecordNotFoundException {
+        Optional<Book> getBook = bookManagerRepository.findById(bookId);
+        if (!getBook.isPresent())
+            throw new RecordNotFoundException("Delete Failed !!! Book with id : " + bookId + " not Found in DB !!!");
         bookManagerRepository.deleteById(bookId);
     }
